+++ util/storage-samples/storage-owner-sample/src/main/java/org/fido/iot/storage/OwnerDbManager.java
@@ -4,6 +4,7 @@
 package org.fido.iot.storage;
 
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.sql.Connection;
@@ -15,12 +16,15 @@ import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Calendar;
+import java.util.List;
 import java.util.UUID;
 import java.util.stream.Stream;
 import javax.sql.DataSource;
 import org.fido.iot.protocol.Composite;
 import org.fido.iot.protocol.Const;
+import org.fido.iot.protocol.CryptoService;
 import org.fido.iot.protocol.RendezvousInfoDecoder;
+import org.fido.iot.serviceinfo.SdoWget;
 
 /**
  * Owner Database Manager.
@@ -403,6 +407,13 @@ public class OwnerDbManager {
    */
   public void addDeviceTypeOwnerSviString(DataSource ds, String deviceType, String ownerSviString) {
 
+    if (!ownerSviString.equals("null")) {
+      ownerSviString = insertModActivateSviEntry(ownerSviString);
+      if (checkWgetInSviString(ownerSviString) && isWgetVerificationEnabled(ds)) {
+        ownerSviString = insertWgetContentHash(ds, ownerSviString);
+      }
+    }
+
     if (checkSviIdReferentialIntegrity(ds, ownerSviString)) {
       String sql = "MERGE INTO DEVICE_TYPE_OWNERSVI_STRING VALUES (?,?);";
       try (Connection conn = ds.getConnection();
@@ -419,6 +430,82 @@ public class OwnerDbManager {
   }
 
   /**
+   * For every module change in sviString, inserts mod:active SVI entry.
+   *
+   * @param sviString svi mapping string to be updated
+   * @return updated sviString with activate module entries
+   */
+  public String insertModActivateSviEntry(String sviString) {
+    StringBuilder sb = new StringBuilder();
+    List<String> sviEntries = Arrays.asList(sviString.split(","));
+    List<String> modSequence = new ArrayList<>();
+
+    for (String sviEntry : sviEntries) {
+      String[] sviEntryArray = sviEntry.split(SVI_ENTRY_DELIMETER);
+      String[] modMsgDelimeted = sviEntryArray[0].split(SVI_MODMSG_DELIMETER);
+      modSequence.add(modMsgDelimeted[0]);
+    }
+
+    sb.append(modSequence.get(0));
+    sb.append(":active=activate_mod,");
+    sb.append(sviEntries.get(0) + ",");
+    for (int i = 1; i < modSequence.size(); i++) {
+      if (!(modSequence.get(i - 1).equals(modSequence.get(i)))) {
+        sb.append(modSequence.get(i));
+        sb.append(":active=activate_mod,");
+      }
+      sb.append(sviEntries.get(i) + ",");
+    }
+
+    sb.deleteCharAt(sb.length() - 1);
+    return sb.toString();
+  }
+
+  /**
+   * Determines if wget module is included in svi.
+   *
+   * @param sviString svi mapping string
+   * @return checks if wget module entries are present in the service info string
+   */
+  public boolean checkWgetInSviString(String sviString) {
+
+    List<String> sviEntries = Arrays.asList(sviString.split(","));
+
+    for (String sviEntry : sviEntries) {
+      String[] sviEntryArray = sviEntry.split(SVI_ENTRY_DELIMETER);
+      String[] modMsgDelimeted = sviEntryArray[0].split(SVI_MODMSG_DELIMETER);
+      if (modMsgDelimeted[0].equals(SdoWget.NAME)) {
+        return true;
+      }
+    }
+
+    return false;
+  }
+
+  /**
+   * Checks if content verification is enabled for wget file content.
+   *
+   * @return checks wget verification enable status
+   */
+  public boolean isWgetVerificationEnabled(DataSource ds) {
+    boolean isWgetContentVerificationEnabled = false;
+
+    String sql = "SELECT WGET_SVI_MOD_VERIFICATION FROM TO2_SETTINGS WHERE ID = 1";
+    try (Connection conn = ds.getConnection();
+        PreparedStatement pstmt = conn.prepareStatement(sql)) {
+      try (ResultSet rs = pstmt.executeQuery()) {
+        if (rs.next()) {
+          isWgetContentVerificationEnabled = rs.getBoolean(1);
+        }
+      }
+    } catch (SQLException e) {
+      throw new RuntimeException(e);
+    }
+
+    return isWgetContentVerificationEnabled;
+  }
+
+  /**
    * Ensures the referential integrity between OWNER_SERVICEINFO and DEVICE_TYPE_OWNERSVI_STRING
    * tables.
    *
@@ -428,6 +515,11 @@ public class OwnerDbManager {
    * @param ownerSviString owner serviceinfo string
    */
   private boolean checkSviIdReferentialIntegrity(DataSource ds, String ownerSviString) {
+
+    if (ownerSviString.equals("null")) {
+      return true;
+    }
+
     ArrayList<String> expectedSviIds = new ArrayList<>();
     ArrayList<String> actualSviIds = new ArrayList<>();
 
@@ -475,6 +567,86 @@ public class OwnerDbManager {
     } catch (SQLException e) {
       throw new RuntimeException(e);
     }
+  }
+
+  /**
+   * Calculated sha384 hash of file content to be transferred using wget:url. Adds the hash entry to
+   * OWNER_SERVICEINFO table Includes appropriate mapping in the sviString for corresponding hash
+   * file.
+   *
+   * @param sviString svi mapping string to be updated
+   * @return updated sviString with activate module entries
+   */
+  public String insertWgetContentHash(DataSource ds, String sviString) {
+    StringBuilder sb = new StringBuilder();
+    List<String> sviEntries = Arrays.asList(sviString.split(","));
+    List<String> newSviEntries = new ArrayList<>();
+
+    for (String sviEntry : sviEntries) {
+      String[] sviEntryArray = sviEntry.split(SVI_ENTRY_DELIMETER);
+
+      // Get the filename whose hash needs to be calculated.
+      if (sviEntryArray[0]
+          .trim()
+          .equals(SdoWget.NAME + SVI_MODMSG_DELIMETER + SdoWget.KEY_FILENAME)) {
+        String filename = null;
+        String sviId = null;
+        byte[] sviContent = null;
+
+        String sql = "SELECT CONTENT FROM OWNER_SERVICEINFO WHERE SVI_ID = ?";
+        try (Connection conn = ds.getConnection();
+            PreparedStatement pstmt = conn.prepareStatement(sql)) {
+          pstmt.setString(1, sviEntryArray[1]);
+          try (ResultSet rs = pstmt.executeQuery()) {
+            if (rs.next()) {
+              filename = new String(rs.getBytes(1), StandardCharsets.UTF_8);
+              // concatenate the filename and the hash to generate a unique SVI_ID.
+              sviId = new StringBuilder(filename + "_hash").toString();
+            }
+          }
+        } catch (SQLException e) {
+          throw new RuntimeException(e);
+        }
+
+        // Read the file content whose hash needs to be calculated.
+        sql = "SELECT CONTENT FROM OWNER_SERVICEINFO WHERE SVI_ID = ?";
+        try (Connection conn = ds.getConnection();
+            PreparedStatement pstmt = conn.prepareStatement(sql)) {
+          pstmt.setString(1, filename);
+          try (ResultSet rs = pstmt.executeQuery()) {
+            if (rs.next()) {
+              sviContent = rs.getBytes(1);
+            }
+          }
+        } catch (SQLException e) {
+          throw new RuntimeException(e);
+        }
+
+        // calculate SHA384 hash of the file content.
+        byte[] wgetFileContentHash =
+            (new CryptoService().hash(Const.SHA_384, sviContent)).getAsBytes(Const.HASH);
+
+        // Adds the hash entry to OWNER_SERVICEINFO table.
+        addServiceInfo(ds, sviId, wgetFileContentHash);
+
+        // add the mapping to GUID_OWNERSVI
+        newSviEntries.add(
+            (SdoWget.NAME + SVI_MODMSG_DELIMETER + SdoWget.KEY_SHA + SVI_ENTRY_DELIMETER + sviId));
+      }
+    }
+
+    int i = 0;
+    for (String sviEntry : sviEntries) {
+      sb.append(sviEntry + ",");
+      String[] sviEntryArray = sviEntry.split(SVI_ENTRY_DELIMETER);
+      if (sviEntryArray[0].trim().equals(SdoWget.NAME + SVI_MODMSG_DELIMETER + SdoWget.KEY_URL)) {
+        sb.append(newSviEntries.get(i) + ",");
+        i++;
+      }
+    }
+
+    sb.deleteCharAt(sb.length() - 1);
+    return sb.toString();
   }
 
   /**
